Correct the instructions and the run script
Trying out the instructions myself, I found two little bugs that are
corrected with this commit.  I also reduced the default number of CPU
cores from 16 to 4, so that the setup can be run locally on a personal
computer, without the need for a server or a larger desktop computer.